refactor: SARIF 파일 파싱 없이 URL만 DB에 저장하도록 구조 변경
