+++ src/application/code-analysis/start-code-analysis.ts
@@ -10,11 +10,19 @@ export async function startCodeAnalysis(versionId: number) {
 
   if (!version) throw new Error('Version not found');
 
+  // 상태 업데이트
   await updateVersionStatusSafely(versionId, {
     codeStatus: 'pending',
     flowStatus: 'pending',
   });
 
+  await prisma.codeAnalysis.upsert({
+    where: { versionId },
+    update: { status: 'pending' },
+    create: { versionId, status: 'pending' },
+  });
+
+  // 워크플로우 트리거
   await triggerSemgrepWorkflow({
     versionId,
     repoUrl: version.project.githubUrl,
